calculateScenes moved to create-video
